Reading in a single packet.
This reads in a single packet.

At this point, it needs more error checking, as this reads in the header
and associated packet. It will be the most used function in the entire
applications.

So, error checking and testing. For this, I'm going to need some way of
testing against known good data. Can I fake this with strings, or do I
need a bunch of test files?